+++ eurorack/src/system/audio.rs
@@ -1,6 +1,9 @@
 use daisy::audio::{self, Block, Interface};
 
-pub const SAMPLE_RATE: u32 = audio::FS.to_Hz();
+// pub const SAMPLE_RATE: u32 = audio::FS.to_Hz();
+// NOTE: The SAMPLE_RATE needs to be adjusted. Probably because the clock on
+// STM32 is unable to exactly match the speed.
+pub const SAMPLE_RATE: u32 = 47_810;
 pub const BLOCK_LENGTH: usize = audio::BLOCK_LENGTH;
 
 static mut BUFFER: [(f32, f32); BLOCK_LENGTH] = [(0.0, 0.0); BLOCK_LENGTH];